db: modify long-lat type on initdb script

+++ db/sql/initdb-table.sql
@@ -3,8 +3,8 @@ CREATE TABLE amap (
   amap_id SERIAL NOT NULL,
   amap_name VARCHAR(250) DEFAULT NULL,
   amap_address VARCHAR(2502) DEFAULT NULL,
-  latitude INT NOT NULL,
-  longitude INT NOT NULL,
+  latitude FLOAT(8) NOT NULL,
+  longitude FLOAT(8) NOT NULL,
   geography POINT,
   PRIMARY KEY(amap_id)
 );
